Added Solution Of Problem 1475B.

+++ Rating 900/VasilijeinCacak_1878C.java
@@ -5,9 +5,9 @@ public class VasilijeinCacak_1878C{
         Scanner sc=new Scanner(System.in);
         int t=sc.nextInt();
         while(t-->0){
-            long n=sc.nextInt();
-            long k=sc.nextInt();
-            long x=sc.nextInt();
+            long n=sc.nextLong();
+            long k=sc.nextLong();
+            long x=sc.nextLong();
 
             long minSum=k*(k+1)/2;
             long maxSum=k*(2*n-k+1)/2;
